Software-Recommendation-System/requirements.txt
flask==2.2.3
flask-cors==3.0.10
pandas==1.5.3
scikit-learn==1.2.2
numpy==1.24.2
google-generativeai==0.3.1 